Update provisioning-skype-room-system-accounts-in-office-365.md
I think that the product behavior must have changed, because the system won't allow us to run enable-csmeetingroom on an account that is not licensed for Skype for Business Online. We have had multiple support cases on this. The most recent one is: 7976962. 
Also the workflow presented on the diagram should be changed: The step "Enable account in SfB Pool" should be after "Assign SfB online license via control panel".