Added info on supported Hybrid topology for SIP domains
Added a Note regarding the need of having all SIP domains online and Onprem to avoid IM and presence issues. This is very important because still today, we have customers not following this:

For a Hybrid environment to work correctly, every SIP domain that exists in your Onpremises environment needs to also exist in your Office 365 tenant and vice-versa. You cannot have some SIP domains online only and some domains Onprem only, this is not a supported topology in Skype for Business. Otherwise, Presence, IM and other features will not work correctly.