Update SRSv2 installation documentation
The installation script now automatically handles downloading any
Windows updates that may be required. The manual steps for this are
obsolete, and have been removed.

Language packs must still be manually downloaded and placed. The script
does not, however, indicate where downloaded LPs should be placed.
Updated the documentation to clarify this, pending the built-in
documentation being fixed.