Update country-and-region-availability-for-audio-conferencing-and-calling-plans.md
This document should be updated now that the introductory period for dial out to 44 countries has expired as of 6/30/2018. 
https://docs.microsoft.com/en-us/skypeforbusiness/legal-and-regulatory/complimentary-dial-out-period. 

Suggest removing note about complimentary period. 
Also paragraph starting on Line 154 refers to list below, but there is no list below and I believe this related to complimentary countries which is no longer available. We need clarification on when dial out requires communication credits. Is this whenever calling outside of country/region user is assigned to in O365?


"Dialing out from a meeting to another country/region in the world that is not listed below is available using [Office 365 Communication Credits](../skype-for-business-and-microsoft-teams-add-on-licensing/what-are-communications-credits.md). For those users, you will need to [set up Communications Credits for your organization](../skype-for-business-and-microsoft-teams-add-on-licensing/set-up-communications-credits-for-your-organization.md)."