Update let-your-teams-users-communicate-with-other-people.md
I tried to update this but I'm not sure of the intent of the Skype for Business Online Organization section. Couple of things:
1. Teams and Skype are now managed under one section in the Modern Portal - should be available to all customers shortly if not there now.
2. The configuration for Teams and Skype federation online is identical - there is no difference
3. The default is allow all domains. Once any domain is added to the list its no longer open federation. That said the modern portal has a bug right now that is showing Microsoft managed domains. 
https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fdomoreexp.visualstudio.com%2Fad4dc0b3-25d1-44f6-a766-8670592cc9a9%2F_workitems%2Fedit%2F360017%3Ftracking_data%3DeyJTb3VyY2UiOiJFbWFpbCIsIlR5cGUiOiJOb3RpZmljYXRpb24iLCJTSUQiOiIxOTU3ODAiLCJTVHlwZSI6IlBFUiIsIlJlY2lwIjoxLCJFbGVtZW50Ijoic3VtbWFyeV9pZF9saW5rIn0%253d&data=02%7C01%7C%7Cf8703c83fa5440e984c108d60dc6d944%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636711545032522533&sdata=pbirTWnDyeISrcsoRO8TX1j945w4IVDNXvNNPVsvBtw%3D&reserved=0

This document should be changed to just one section managing Teams/Skype federation with Enabled set for default to allow all organizations and if administrator wants to restrict then they should add specific domains. If other organizations they want to communicate with also restrict domains, they should ask that administrator to add their domain to their tenant allow list.