Update Known-issues.md
Removing
|**Issue title**|**Behavior / Symptom**|**Known workaround**|**Discovery date**|
|:-----|:-----|:-----|:-----|
|Skype for Business on-premises users aren't getting my messages  <br/> |Messages will not complete when Microsoft Teams users try to send a message to another individual who is using Skype for Business on-premises.  <br/> | Interoperability between Teams and users hosted on Skype for Business Online is supported. Teams users can send 1:1 chats to non-Teams users using Skype for Business Online. <br/> Interoperability between Teams and users hosted on Skype for Business on-premises is not supported. Teams users cannot send 1:1 chats to non-Teams users using Skype for Business on-premises.  <br/> |11/2/16  <br/> |

This is not correct. Interop is possible between Skype for Business on premises and Teams online. This is covered in this document: 
https://docs.microsoft.com/en-us/microsoftteams/migration-interop-guidance-for-teams-with-skype

Please check with BJ Whalen if you have any questions or concerns. 

Removed redundant and incorrect section about Missing Meetings. 
Corrected remaining section. We do show the meeting app even for on premise builds < Exchange 2016 CU3. Now it shows but with error that calendar details couldn't be retrieved.