Update overview.md
update to include announcement of Certified for Microsoft Teams certification.  There will be coinciding edits to USB devices, IP phones, and meeting room devices.

also updated "applies to" to include Teams.


question to reviewer: i want the "new" section to be in a blue box like the "note" section currently on 
IP phone pages, but i want the text to say new instead of note. it seemed that this may be the way to do it, but i also saw documentation which indicated we may only have option for 4 specific types of note blocks. If i can't use new, i'd prefer "Important" instead of "Note"