Update direct-routing-sbc-multiple-tenants.md
To activate a user in a with sip address matching sip domain of SBC base FQDN requires the user is licensed for SfB Plan2 today, and Teams in the future when we change licensing sku's. To avoid that detail just added requirement for E1, E3, or E5 license. Not 100% sure if these are the only licenses today that include SfB Plan2 and in future will auto provision SfB shadow accounts with Teams license.