Update configuring-teams-calling-quickstartguide.md
Updating for changes in Calls App which now shows if 1:1 calling is enabled and currently without voicemail. 

We're getting lots of questions from customers on change in behavior so asking this gets updated ASAP. Thanks. 

Was tempted to make changes that will be occurring 11/15/2018 for deprecation of Interop policy but figure this is being tracked for when those changes occur.