Update direct-routing-user-accounts-in-a-hybrid-environment.md
Updated supportability statement for the hybrid environment. We had redundant information. We don't need to create users in on premise pool and then move. What we must do is enable the users with on premise management tools with hosting provider of online so that on premise is aware of the sip enabled user and their service location.