Sample Script to Automate Firewall Rule Creation
We've been getting valid feedback from administrators that they do not want to ask end users to cancel security dialog for Firewall rule with explanation that it will work OK. This is both confusing to end users as well as bad practice to instruct users to ignore security dialogs. Administrators are asking us to please provide a mechanism to allow them to create appropriate firewall rules that will avoid the dialog. 

We have received the following sample script from engineering that works great and avoids the prompt. This script can be run as a GPO computer startup script. Can we add this sample script to this documentation so that administrators know how to control this end user experience.

Sample Script

$users = Get-Childitem c:\users
foreach ($user in $users) 
{
    $Path = "c:\users\" + $user.Name + "\appdata\local\Microsoft\Teams\Current\Teams.exe"
    if (Test-Path $Path) 
	{
	        $name = "teams.exe " + $user.Name
	        if (!(Get-NetFirewallRule -DisplayName $name))
		{
	            New-NetFirewallRule -DisplayName “teams.exe” -Direction Inbound -Profile Domain -Program $Path -Action Allow
		}
	}
}