Update Known-issues.md
Hi,
I purpose a change to the line #161 as our internal information mentions both Strings actually. This is also confirmed by the Teams EEE Roman Maddox.

Additionally, on the powershell cmdlet that was added before, it would in fact replace the ENTIRE EWSAllowList, with just *skypespaces* this can be very very critical to the customer's environment as we won't have a way to recover what he had before!

The cmdlet I added: Set-organizationconfig -EwsAllowList @{Add="*MicrosoftNinja*","*SkypeSpaces*"} ;
Is effectively just adding those 2 strings, instead of replacing in its entirety.

Note: this type of issue is arising more and more, and we're getting multiple tickets which lead to confusion when we discuss it with the customers and direct them here, where only *skypespaces* is mentioned (which won't fix the meeting access issue).