Update get-clients.md
Updated the PowerShell script. I know it says it's a sample, but why not present a more robust sample? :)
Added synopsis and a bit more robustness. Specifies requires PS version 3, excludes local admins and Public users, looks for a firewall rule containing the path to Teams.exe for the users instead of a firewall rule name. Specifies TCP and UDP port instead of allowing all ports.