Update teams-channels-overview.md
Updated 2500 members to 5000, as per https://github.com/MicrosoftDocs/OfficeDocs-SkypeForBusiness/issues/1575

One point to note, MikeParker365, and to make it easier for the editor is to to keep with the terminology used in the document when reporting errors or amendments. In this instance, it's straightforward as though [unlimited] users could be invited there's only a finite number of members (if that makes sense). Thanks, Mike.