Update Exchange-Teams-interact.md
Changing Table for features:
Changing Profile Photo for Exchange On Premises is same as for Calendar access. If EX2016 CU3+ it works.
Changing Voicemail support to yes + supplemental Note 5 as PM didn't like the No since customers were than thinking we didn't support voicemail for Teams if Exchange is on premise.
Changing support for Connectors and Tabs to Yes for on premise Exchange. We do support Connectors and adding TABS if Exchange is on premise. I confirmed this in lab for Exchange 2013 on premise users. 

Changing the configuration of Oauth requirement if on premise to fix syntax error and be more explicit about requirement for the the new Exchange Oauth protocol.