Change bold font to contact email support
The answer to user question might not be available to the public because for the devices to be certified and be part of SBC, MS needs to perform an intense test.  But it was indicated on the article: " If you have any questions about SBC certification program for Direct Routing please email to drsbccertification@microsoft.com" Can we instead direct the user to email the support or perhaps a uservoice? Related Uservoice: https://www.skypefeedback.com/forums/299910-recently-released/suggestions/13732851-expand-skype-certified-devices.