Update information-barriers-in-teams.md
Noticed that issue 2146 was related to this doc only after I had done PR's for both. Had first done PR for 2144, then did PR for 2146 and realised it's same document and that wording must be changed from 'private' to 'public' for this one.  This issue raiser had it as private, but DanMHyre raised issue in 2146 so added commit to this one to make sure it's reading public in this PR.