Remove unnecessary text
The article introduction includes unnecessary text, which actually works against the purpose of the article. For example, expressions such as "contains recommendations" and "temporarily apply these procedures" contravene with "you must" in the first paragraph after the removed ones. Also, the article should not "contain information that shows how to help lower security settings" and it does not show "how to temporarily turn off security features on a computer". The article includes only the list of exclusions that are required to ensure that the antivirus scanner does not interfere with the operation of Skype for Business Server. The Product Group considers them as required and expects them to be enforced before engaging on an issue. With the current introduction, it is difficult to confirm the customer they are required and not optional.