Update limits-specifications-teams.md
Add this to Meeting Limits
|Max PowerPoint File Size | 2GB|

But please verify with PG that this is correct. We are obtaining this information from here: https://support.office.com/en-us/article/how-certain-features-behave-in-web-based-powerpoint-a931f0c8-1305-4428-8f7c-9cfa00ef28c5#OfficeVersion=PowerPoint_for_the_web

And it's our understanding that this limit applies to Teams as we use the same viewer as PPT online.