Decomission of diagnostic alerting
I needed to remove all references to the diagnostic alerting and this feature no longer exists.  I also made some edits to conform with CQDv3's GA at ignite: Updated prerequisites and removed reference to PII.  This will be the first of a couple edits, however we needed to make some changes quickly to conform to Ignite guidance.