Note on measuring network performance through ICMP
PG is discontinuing ICMP requests to 13.107.8.2 as that is not an effective way of measuring it. Some customers have been unable to use it already, and this is creating some Service Requests to CSS. I've added a note.
See https://icm.ad.msft.net/imp/v3/incidents/details/157254636/home for more info from Mykhailo Moroz.