Changed description on Override caller ID Policy
The subtract I've changed of the text was confusing, as enabling "Override Caller ID Policy" will allow only the user to choose if he wants to call showing his number or as anonymous. It won't allow the user to change if he wants to see the number of the caller or not.