Update create-a-phone-system-call-queue.md
Proposing change since the tested behaviour verified is that even if the agent is offline or set to do not disturb, although the call does not notify the user, it is still routed to that user and fail to connect since the user is unavailable. There have been support requests about this matter. For the moment the call routing for call queues is not presence based. This can be the case in the future but there is no ETA for that to happen at this moment.