Update devices-usb-devices.md
Hi @dstrome ,
As we had discussed in the meeting, this is the first set of changes we would like to get on to the docs page. Could we please have it published by 3/17/2020. I have updated the docs page at line 33 updated the month and date of update, Line 36 added a new column. Lines 34- 85 added the certified devices in the year 2020 and 2019. .
Regards
Bhavani