Update devices-usb-devices.md
Changed the update date to march 30,2020 at line 33 , updated the line 37- 43, with the recently certified devices  and updated the table with all the missing old certified device.
@dstrome - As I am not able to see the preview of the content, i would request you to kindly take a look at the format of the page before merging it.
This is the 3rd phase and the last phase of getting the old devices updated on this Peripherals page. here on I would create a PR once there are any new certified devices.