Update Exchange-Teams-interact.md
Adding the additional steps required to allow for scheduling Teams calendar meetings on behalf of another user in Outlook. Reviewed these steps with Engineering and confirmed in the lab with in depth log analysis. The Teams scheduling application requires permissions to check the delegators mailbox for delegate permissions. The application is using the Skype for Business Online Application ID. To enable these permissions, the Exchange server needs to configure a Partner Application for the Skype Online Application ID with a linked account that has the Exchange UserApplication role assigned. 

With COVID 19 usage increases we've seen a large number of cases where delegation is not working due to these missing steps which were never documented. 

@LolaJacobsen @kenwith PM for this feature is ashwin.appiah@microsoft.com