Update Exchange-Teams-interact.md
Scheduling a meeting with Outlook and the Teams add-in for the user itself is supported without EWS/OAuth etc. When it comes to delegation it's not possible to schedule a meeting on behalf without some basic requirements. The lack of information in our public docs leads into widely confusion in our customers. Espacially as a result of several C-19 Teams rollouts. Please add a hint here. THX

https://dev.azure.com/Supportability/UC/_wiki/wikis/UC.wiki/3769/Teams-Delegate-Meeting-Requirements-and-Limitations