Update set-up-phone-system-voicemail.md
@CarolynRowe cc @roykuntz 
Carolyn I discussed these changes with Roy. 

We need to add a reference back to the Section on configuring CVM with Exchange Server in this Doc https://docs.microsoft.com/en-us/skypeforbusiness/hybrid/plan-cloud-voicemail. Under the requirements for Exchange UM on-premises. We only refer to how to configure SfB, but not how to configure Exchange for CVM. Since we do here, we should cross reference back to here. 

cc: @dstrome