Update meeting-policies-in-teams.md
@kenwith @LolaJacobsen @sonu-arora 
CC @revaldiv who worked with me on this.

Changes made to simplify, clarify and correct meeting join experience based on meeting policy settings and meeting options as well as to add information about the fact that dial in users will land in the lobby if unauthenticated users cannot start meetings and they are the 1st to join the meeting. This is true even when Dial In users are set to bypass lobby. This requires a user from the organization to join with a Teams client to admit them as there is currently no way to do this via dial in meeting controls. The only dial in meeting control today is mute. 

The document as written mixed the teams client controls with dial in controls. They are separate and mutually exclusive.