[DO NOT MERGE] SEO edits for Overview of security and compliance
Edited the description to use the most searched for terms.

Per the SEO expert, added addtional links to the Privacy section. Links are for statements about how Microsoft protects privacy for consumers and IT pros.

Also edited the Compliance standards section to place the HIPAA compliance in it's own sentence, as this is a common search term by itself.