[DO NOT MERGE] SEO edits for Upgrade to Teams on-premises
Updated the description and H1 per the SEO expert's recommendations. Added the "admin" part to the description since it's called out specifically in the H1.

The H1 change is based on "migrate" being the highest hit search term, but "upgrade" might be preferred as a marketing tool.  Another option might be to use "Upgrade and migrate...".

Removed a few extra spaces I noticed in the Overview.