Update direct-routing-protocols-sip.md
@CarolynRowe @FilippSe 

Fix for Contact header which requires Port. 
From: <sip: 68712781@sbc1.adatum.biz;transport=tls>, 
To: <sip: 68712781@sbc1.adatum.biz:5058;transport=tls>

Also refined information about contact header versus record route due to changes and requirements for SBC Proxy. What it says now

According to RFC 3261, Contact header is required in any request that can result in a new dialog. The Record-Route is only required if a proxy wants to stay on the path of future requests in a dialog. If a proxy SBC is in use with [Local Media Optimization for Direct Routing] (https://docs.microsoft.com/en-us/MicrosoftTeams/direct-routing-media-optimization), a record route will need to be configured as the proxy SBC needs to stay in the route.

Microsoft recommends using only Contact header if a proxy SBC is not used:

Per RFC 3261, Record-Route is used if a proxy wants to stay on the path of future requests in a dialog, which is not essential if no proxy SBC is configured as all traffic goes between the Microsoft SIP proxy and the paired SBC. 

The Microsoft SIP proxy uses only Contact header (not Record-Route) to determine the next hop when sending outbound ping Options. Configuring only one parameter (Contact) instead of two (Contact and Record-Route) simplifies the administration if a proxy SBC is not in use.