Update meeting-settings-in-teams.md
@LanaChin @SerdarSoysal  @sonu-arora 

Updating this documentation to reflect changes made by engineering per this ICM: https://portal.microsofticm.com/imp/v3/incidents/details/195655177/home

The original plan was that enabling DSCP tagging would cause destination ports of 3479-3481 to be used for Audio, Video and Desktop Sharing but apparently this caused too many failures because of customer firewall rules and we are now sending all traffic to UDP Port 3478. However, customers can requests to have an exception made to their tenant and their tenant ID will be added to ECS list to allow port segregation. 

From IC 

Worth noting that we wanted to make it so relays always behave as customer asks. We found that we could not because it breaks a bunch of customers who have only opened 3478 in their networks. So we still have this behavior as an ECS key that we can set, if the customer really wants it. Using source ports for QoS marking is generally the better approach though.