Update Exchange-Teams-interact.md
@dstrome 

David it's endless :) I had previously made change to the table for profile photo support and I guess missed the paragraph at the start that states ability to change photo (or that was added after the fact). We no longer support changing profile photo because method changed from MT EWS call to Graph Explorer call which requires REST API support for the Exchange Mailbox. From my testing I find that a 404 is returned when attempting to save the photo and user gets and error message stating that there was a problem saving the photo please try again. We require some work from engineering here too because we shouldn't show an option to change if not supported. For now, I wanted to get this article consistent with the current experience and have the table and text align. 

Do you know who owns this support topic which is also now wrong on this point: https://support.microsoft.com/en-us/office/change-your-picture-in-teams-7a711943-9248-420e-b814-c071aa8d9b9c#ID0EAABAAA=Desktop