Update upgrade-to-Teams-on-prem-overview.md
clarified that onprem users can't be TeamsOnly and that using Teams in Islands mode is not same as TeamsOnly.

btw, for the "use Teams in Islands mode" I added a link to the Migraiton-Interop-guidance-forTeams-with-skype for more details. But i think a better link would be this specific subsection on the following page:

https://docs.microsoft.com/en-us/MicrosoftTeams/upgrade-to-teams-on-prem-upgrade-methods#overlapping-capabilities-method-using-islands-mode

i wasn't sure how to link to a specific section in a page.  If you can fix that, pls do so.