Make both callout/Important boxes re Live Events the same
Prior PR updated the callout box to have caveat sentence re the assessment/triage stage of requests that go to the Live Events Assistance Team.  But that update was only to one box, and there are 2 with substantively the same language.  So this PR copies the updated box and its less-wordy text, and including the caveat sentence.