Update audio-conferencing-on-network.md
@CarolynRowe 
Updated to clarify that this feature is only supported in Teams Meetings. It does not require Teams Only client, it just requires Teams Meetings. Calls are initiated via the Online Teams Meeting service, so mode of Teams client is not relevant. 

Also added clarity that outbound calling is destined to on-network destination because its not supported to route these calls beyond corporate network to PSTN.