Update create-a-phone-system-auto-attendant.md
@MikePlumleyMSFT 
PM for this is naveenv@microsoft.com 

Added information about how to configure the destination number because we have cases on this coming in. Found this in TAP and had requested UI warning to prevent customer errors. 

Also revised wording for the Skype for Business Hybrid to make it clearer. Customers and support engineers were not clear on this point. If a tenant is Skype for Business Hybrid and Skype users on premise will call AA, then routing will fail if External Phone number is used for routing to PSTN. In this case customers needs to use a target of a user that is on-premises EV enabled with forwarding to PSTN.