Update cloud-recording.md
Fixed the text around Stream recordings. We had it saying that Stream recordings were deleted after 21 days. That's totally not correct. 
Also fixed the wording in the paragraph so that it reflects that fact that we've already enabled the agility to store recordings outside of the country and in the nearest regional Stream DB.