Update collab-bar-deploy.md
Rename from "collaboration bars for Microsoft Teams" to "Microsoft Teams Rooms". Microsoft Teams Rooms on Android is not an official name. Rather, it is only used - when necessary - to differentiate between unique features of the Windows and Android platforms. As such, "on Android" is used sparingly -primarily at the beginning of the article to set the scope of the document.

Document cleaned up a little too. Some parts got a little wordy.