Update direct-routing-protocols-sip.md
There is an inconsistency on how our customers configure SIP header for OPTIONS and INVITE messages, mostly ignoring the need of using FQDNs instead of IPv4 addresses. The changes proposed address multiple scenarios found in customer and DRaaS provider's scenarios, which per se can affect not only call establishment but for OPTIONS failing, messages, a downtime on SBC communication will be experienced. The example provided adds clarity and consistency to the SIP headers manipulation rules that need to be in place on SBC. This will also assist support on driving SBC providers to update the documentation to match the Microsoft's recommendations and requirements.