Update teams-panels.md
I'm all for QoS applied to all Teams networks. However, stating that QoS is a preferred Site readiness requirement implies that Teams Panels send media. At the current release level, they do not - and may not ever. QoS is not needed for AD sign in, EWS lookups, etc.

You can sign in to a Teams panel without a Teams Rooms Standard license. I just tested with an E5. Therefore, it appears we are making an incorrect assumption that the Teams Rooms Standard license will necessarily be assigned to the resource account.

Many organizations apply extra E3 licenses or whatever to resource accounts as they have extras and don't want to buy more licenses. And you can use the panel without a Teams rooms Standard license. So we should strongly encourage the Teams Rooms license but (at least as of right now) that license is not a requirement for a panel to work.

Teams Rooms standard does *not* include a resource account license. Resource accounts licenses are free and as such do not require a license. 

As for the Naming Convention tip.....in most cases, the resource account will already be set up using a different naming convention, e.g. "CR-Aetna" for conference room Aetna. I believe in most cases, a resource account will not be made for the Panel as it will already exist, even in a non-Teams environment. As such, the tip, while useful, can be confusing. 

This is all new and I could be wrong on all of this :) I'm still learning but this is my feedback on the article. Hope I'm not coming off snarky or anything. I truly appreciate that this article is out there and have already shared it on social media to help raise awareness.