Clearer PowerShell usage guidance for view-only experience
As the PowerShell snippet provided did not clarify which module to use, I've added further detail to explain this. The `MicrosoftTeams` module appears to lack any docs.microsoft.com documentation, so only a PowerShell Gallery link has been able to be provided, or the legacy link to the `SkypeForBusiness` module's documentation. I recommend whoever from Microsoft picks this up checks why there is no documentation for the `MicrosoftTeams` module; all hits for the cmdlet listed in this page bring up the legacy `SkypeForBusiness` module, which even itself talks about it having been deprecated with no links to the replacement module...