Update hardware-requirements-for-the-teams-app.md
At line 38 added some suggested language re OS and older builds.  I am not exactly sure what to say - but we want to include something like Windows 10 and 1803 is not recommended.  See ICM: 236372082  where cx video still on bc leave of a issue and tied to OS.  We want something similar to what Mac is saying on line 53 for supported versions and  how we say similar for mobile clients on line 90.