Add recommendation for 32/64-bit version
I would suggest to add a recommendation for the 64-bit version of Teams on Windows. If you install the 32-bit version, a banner is displayed in the Teams client telling the user that it is recommended to use the 64-bit version. So we should also add this recommendation to this Docs article.